http: Check for PKCS#11 URI in ssl_cert and set
Curl can use the PKCS#11 URI for certificates but the type has to be
set to "ENG" for it to do so.

Curl it self detects it by checking the cert for pkcs11: and set the
type to engine implicitly.
This patch adapts the change the same change for git.

Signed-off-by: Björn Bidar <bjorn.bidar@jolla.com>